Retire all the Bison/Flex stuff.
It'll be better to write the lexer and parser manually. Diving into
Bison and Flex was really bringing up a ton of extra overhead and
really only promoted a better understanding of the Bison/Flex
environment. One of the major goals of this project is a better
understanding of the lex/parse process.