Tons of lexer changes. Actually make use of the lexer.
The lexer here now properly looks along the char stream.
Only a handful of things are tokenized right now and the
values of things like character constants and string
literals are not yet saved into the Token object. I
think I'm going to want to create a buffer object or
something to hold the values.